refactor: split exerciseImporCSV into custom hook
